+++ lib/screens/passwords_overview_screen.dart
@@ -92,53 +92,101 @@ class _PasswordsOverviewScreenState
             ? Center(
                 child: CircularProgressIndicator(),
               )
-            : Column(
-                children: [
-                  Padding(
-                    padding: EdgeInsets.all(5),
-                    child: Card(
-                      child: Padding(
-                        padding: const EdgeInsets.all(5),
-                        child: TextField(
-                          autocorrect: false,
-                          decoration: InputDecoration(
-                            labelText: tl(context, 'general.search'),
-                            hintText: tl(context, 'general.search_hint'),
-                            suffixIcon: IconButton(
-                              onPressed: () {
-                                _searchTextController.clear();
-                                refreshPasswords(false);
-                              },
-                              icon: Icon(Icons.clear),
+            : Column(children: [
+                Expanded(
+                  child: RefreshIndicator(
+                    onRefresh: () => refreshPasswords(),
+                    child: Scrollbar(
+                      child: Scaffold(
+                        body: CustomScrollView(
+                          slivers: <Widget>[
+                            SliverAppBar(
+                              backgroundColor: Theme.of(context).canvasColor,
+                              flexibleSpace: Container(
+                                padding: EdgeInsets.fromLTRB(13, 3, 13, 0),
+                                child: TextField(
+                                  autofocus: false,
+                                  cursorColor: Theme.of(context).accentColor,
+                                  autocorrect: false,
+                                  style: TextStyle(
+                                    fontSize: 20,
+                                    color: Theme.of(context).accentColor,
+                                  ),
+                                  decoration: InputDecoration(
+                                    border: InputBorder.none,
+                                    focusedBorder: InputBorder.none,
+                                    enabledBorder: InputBorder.none,
+                                    errorBorder: InputBorder.none,
+                                    disabledBorder: InputBorder.none,
+                                    // labelText: tl(context, 'general.search'),
+                                    hintText:
+                                        tl(context, 'general.search_hint'),
+                                    hintStyle: TextStyle(
+                                      color: Theme.of(context)
+                                          .accentColor
+                                          .withAlpha(200),
+                                    ),
+                                    fillColor: Theme.of(context).accentColor,
+                                    focusColor: Theme.of(context).accentColor,
+                                    hoverColor: Theme.of(context).accentColor,
+                                    filled: false,
+                                    suffixIcon: IconButton(
+                                      alignment: Alignment.centerRight,
+                                      onPressed: () {
+                                        _searchTextController.clear();
+                                        refreshPasswords(false);
+                                      },
+                                      icon: Icon(Icons.clear),
+                                      color: Theme.of(context).accentColor,
+                                    ),
+                                  ),
+                                  maxLines: 1,
+                                  controller: _searchTextController,
+                                  keyboardType: TextInputType.text,
+                                  onChanged: _searchPassword,
+                                ),
+                              ),
+                              shadowColor: Theme.of(context).accentColor,
+                              floating: true,
+                              expandedHeight: 30,
+                              forceElevated: true,
                             ),
-                          ),
-                          maxLines: 1,
-                          controller: _searchTextController,
-                          keyboardType: TextInputType.text,
-                          onChanged: _searchPassword,
+                            SliverList(
+                              delegate: SliverChildListDelegate(
+                                [
+                                  SingleChildScrollView(
+                                    child: Column(
+                                      children: [
+                                        Container(
+                                          child: ListView.builder(
+                                            shrinkWrap: true,
+                                            physics:
+                                                NeverScrollableScrollPhysics(),
+                                            cacheExtent: 10000,
+                                            itemCount: passwords.length,
+                                            itemBuilder: (ctx, i) =>
+                                                PasswordListItem(
+                                              passwords[i],
+                                              deletePassword: deletePassword,
+                                              autoFillMode: autofillMode,
+                                              searchQuery:
+                                                  _searchTextController.text,
+                                            ),
+                                          ),
+                                        ),
+                                      ],
+                                    ),
+                                  )
+                                ],
+                              ),
+                            ),
+                          ],
                         ),
                       ),
                     ),
                   ),
-                  Expanded(
-                    child: RefreshIndicator(
-                      onRefresh: () => refreshPasswords(),
-                      child: Scrollbar(
-                        child: ListView.builder(
-                          cacheExtent: 10000,
-                          itemCount: passwords.length,
-                          itemBuilder: (ctx, i) => PasswordListItem(
-                            passwords[i],
-                            deletePassword: deletePassword,
-                            autoFillMode: autofillMode,
-                            searchQuery: _searchTextController.text,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
+                ),
+              ]),
       ),
     );
   }