+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.9.12
+- Dark Theme & Amoled Mode
+- Bugfixes
+
 ## 0.9.11
 - UI/UX improvements
 - Bugfixes